feat: add single-letter shortcuts + internationalization
1) Adds single-letter shortcuts to the main commands.
This is useful for people having difficulties to type, e.g. people with disabilites.
This also allows for compact "functions", using the `do` and `repeat` commands with a string of letters
(e.g. `repeat("GLglGRgr",3)`.
Those also serve as an introduction to loops and can be localized.

The letters command are designed to allow for grid drawing, both on straight and diagonal segments:
- a = small (straight) go, A = large (diagonal) go,
- r/l = small (diagonal) turn, R/L = large (straight) turn.

2) Localize every menus and commands.
Currently with a french translation, but uses gettext (the standard tool for internationalization),
so can be easily ported to other languages.